+++ Backend/Personas/Coder.py
@@ -3,6 +3,7 @@ from typing import List, Dict, Any
 from Personas.BasePersona import BasePersona
 from Personas.PersonaSpecification import CoderSpecification
 from Utilities.ErrorHandler import ErrorHandler
+from Workflows.AutoWorkflow import AutoWorkflow
 from Workflows.ChatWorkflow import ChatWorkflow
 from Workflows.WriteWorkflow import WriteWorkflow
 
@@ -15,6 +16,7 @@ class Coder(BasePersona):
     WORKFLOWS: Dict[str, Dict[str, Any]] = {
         'write': WriteWorkflow(),
         'chat': ChatWorkflow(),
+        'auto': AutoWorkflow()
     }
 
     def __init__(self, name: str) -> None: